feat: add delta time to onUpdate func

+++ src/main/java/io/nova/utils/Time.java
@@ -3,7 +3,7 @@ package io.nova.utils;
 public class Time {
     public static final long applicationStartTime = System.nanoTime();
 
-    public static double getElapsedTimeSinceApplicationStartInSeconds() {
-        return (System.nanoTime() - applicationStartTime) * 1E-9;
+    public static float getElapsedTimeSinceApplicationStartInSeconds() {
+        return (float)((System.nanoTime() - applicationStartTime) * 1E-9);
     }
 }
